+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - Dec 7, 2024
+
+### Added
+
+- Add album url and uri for track results - [Issue](https://github.com/nesaku/SpotifyInfo/issues/4)
+- Add a loading skeleton for the track cover image
+  s
+
+### Changed
+
+- Improve home page styling
+
 ## [3.0.1] - Apr 29, 2024
 
 ### Fixed